+++ src/app/code/community/Shopgate/Cloudapi/Model/Auth.php
@@ -44,10 +44,9 @@ class Shopgate_Cloudapi_Model_Auth
             return $this->retrieveUser(Mage_Api2_Model_Auth_User_Guest::USER_TYPE);
         }
 
-        $store = Mage::app()->getStore($request->getParam('store'));
+        $store      = Mage::app()->getStore($request->getParam('store'));
         $customerId = $this->validateRequest($store);
         if (null !== $customerId) {
-            //todo-sg: maybe pull store from client_id instead? Need to verify that the shop_number store === current?
             $customer = Mage::getModel('customer/customer')->setStore($store)->load($customerId);
 
             return $this->retrieveUser(Mage_Api2_Model_Auth_User_Customer::USER_TYPE, $customer->getId());